Assignment_2_Q2
Assignment 02: Understanding Key API and Software Concepts
Topic	Description
API	API is the medium through which the software or application communicates and interacts. 
SDK	A Software Development Kit provides developers with tools, libraries, and documentation to build applications for specific platforms or frameworks.
Web APIs	Web APIs allows the applications to exchange data using web. 
Library APIs	Library APIs consist of predefined functions and procedures within software libraries, facilitating the integration of specific features into applications.
Prompting and Completion	In simple words, the prompting is input to the llm and completion is response of the LLM.